Mod #170 : move file location
